fix: send_mouse_button 调用 int→uint8_t 隐式转换触发 C4244
循环变量 b 为 int, 传入 uint8_t 参数在 /WX 下报 C4244, 显式转换两处。

+++ lib/ioctl-client.c
@@ -284,7 +284,7 @@ int ioctl_send_mouse(HANDLE hDevice, const mouse_report_t* report) {
     for (int b = 0; b < 5; b++) {
         uint8_t mask = (uint8_t)(1u << b);
         if ((g_PrevMouseButtons ^ now) & mask) {
-            send_mouse_button(b, (now & mask) != 0);
+            send_mouse_button((uint8_t)b, (now & mask) != 0);
         }
     }
 
@@ -328,7 +328,7 @@ int ioctl_reset_devices(HANDLE hDevice) {
     }
     for (int b = 0; b < 5; b++) {
         if (g_PrevMouseButtons & (1u << b)) {
-            send_mouse_button(b, FALSE);
+            send_mouse_button((uint8_t)b, FALSE);
         }
     }
